requirements.txt
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
evdev==1.6.1
Flask==3.0.0
fonttools==4.45.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keyboard==0.13.5
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.25.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pynput==1.7.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-whatsapp-bot==0.0.91
python-xlib==0.33
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sklearn==0.0.post9
threadpoolctl==3.2.0
tzdata==2023.3
urllib3==2.0.4
Werkzeug==3.0.1
